Registers all ePackages that are extended by profiles in the central package registry
This is necessary to fix the issue when having references from
stereotypes to base objects that are dynamic, because when they are
dynamic their metaclass is loaded by the reflective editor and differs
from the one that is referenced in the package registry. When
registering the package of the metaclass, it is resolved from the
registry when loaded in the reflective editor and, hence, the instanceOf
check in EStructuralFeature succeeds.