Fixes issue 35
I fixed this issue by checking in the CreateShortcutAction whether the
imported EPackage normalizes to a platform resource URI and, if yes,
replaces the URI of this resource with its actual nsURI.

I'm not sure if this is actually a very good solution as it may
influence other tools that use that resource instance of the package (as
we change its URI).

The issue that remains is when opening the profile after restarting
Eclipse, an error will occur since the imported package will not be
available anymore. It has to be registered before the profile can be
openend.